Program redirect system
Both the redirect and the timed redirect work now.  A countdown timer is displayed.  The "goNow" button is usually available while the timer is counting down, but if you don't want someone to be able to skip the timer you can include the URL parameter `&goNow=disabled` to remove this button and hide the redirect link.