Double-ensure cancel button works.
`window.close()` is supposed to close the tab, but it sometimes simply doesn't.  For this reason, the tab will first try to go back to the previous page in history, and then try to close itself if there was no previous page, and then, if neither of those worked, the tab will redirect to `about:blank`.